fix: issue sidebar create label form
