+++ apps/app/components/issues/sidebar-select/index.ts
@@ -2,8 +2,9 @@ export * from "./assignee";
 export * from "./blocked";
 export * from "./blocker";
 export * from "./cycle";
+export * from "./estimate";
+export * from "./label";
 export * from "./module";
 export * from "./parent";
 export * from "./priority";
 export * from "./state";
-export * from "./estimate";
